+++ indianapolis-model/src/main/java/com/gdn/entity/PickupDetail.java
@@ -41,6 +41,6 @@ public class PickupDetail {
     private int skuPickupQuantity;
 
     @Column(name = "cbm_pickup_amount")
-    private double cbmPickupAmount;
+    private Float cbmPickupAmount;
 
 }
