+++ indianapolis-model/src/main/java/com/gdn/recommendation/DatabaseQueryCffGoods.java
@@ -11,6 +11,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class DatabaseQueryCffGoods {
     private String id, sku;
-    private Double cbm;
+    private Float cbm;
     private int quantity;
 }
